@@ -1,142 +1,131 @@
 {
+  "name": "pizza-app",
+  "version": "0.0.1",
   "private": true,
+  "main": "node_modules/expo/AppEntry.js",
   "scripts": {
-    "test:e2e": "./bin/downloadExpoApp.sh && detox test --configuration ios.sim.expo",
-    "build:e2e": "detox build -c ios.sim.expo",
-    "ci:test:e2e": "detox test -c ios.sim.release -l verbose --cleanup",
-    "ci:build:e2e": "detox build -c ios.sim.release",
+    "start": "react-native start",
+    "ios": "react-native run-ios",
+    "android": "react-native run-android --active-arch-only",
+    "test:detox": "detox test -c ios.sim.debug",
+    "build:detox": "detox build -c ios.sim.debug",
+    "ci:test:detox": "detox test -c ios.sim.release -l verbose --cleanup",
+    "ci:build:detox": "detox build -c ios.sim.release",
     "compile": "tsc --noEmit -p . --pretty",
-    "format": "npm-run-all format:*",
-    "format:js": "prettier --write '**/*.js'",
-    "format:json": "prettier --write '**/*.json'",
-    "format:md": "prettier --write '**/*.md'",
-    "format:ts": "prettier --write '**/*.ts{,x}'",
-    "hack:types-react-navigation": "rimraf node_modules/@types/react-navigation/node_modules/@types",
-    "hack:types-react-native": "rimraf node_modules/@types/react-native/node_modules/@types",
-    "hack:types-react-test-renderer": "rimraf node_modules/@types/react-test-renderer/node_modules/@types",
-    "lint": "eslint App.js app storybook test --fix --ext .js,.ts,.tsx && yarn format",
+    "format": "prettier --write \"app/**/*.{js,jsx,json,md,ts,tsx}\"",
+    "lint": "eslint index.js App.js app test --fix --ext .js,.ts,.tsx && npm run format",
     "patch": "patch-package",
-    "storybook": "start-storybook -p 9001 -c ./storybook",
     "test": "jest",
+    "test:watch": "jest --watch",
     "adb": "adb reverse tcp:9090 tcp:9090 && adb reverse tcp:3000 tcp:3000 && adb reverse tcp:9001 tcp:9001 && adb reverse tcp:8081 tcp:8081",
-    "prepare": "npm-run-all patch hack:*",
-    "start": "expo start",
-    "android": "expo start --android",
-    "ios": "expo start --ios",
-    "web": "expo start --web",
-    "eject": "expo eject"
+    "postinstall": "node ./bin/postInstall",
+    "bundle:ios": "react-native bundle --entry-file index.js --platform ios --dev false --bundle-output ios/main.jsbundle --assets-dest ios",
+    "bundle:android": "react-native bundle --platform android --dev false --entry-file index.js --bundle-output android/app/src/main/assets/index.android.bundle --assets-dest android/app/src/main/res",
+    "release:ios": "echo 'Not implemented yet: release:ios. Use Xcode. More info: https://reactnative.dev/docs/next/publishing-to-app-store'",
+    "release:android": "cd android && rm -rf app/src/main/res/drawable-* && ./gradlew assembleRelease && cd - && echo 'APK generated in ./android/app/build/outputs/apk/release/app-release.apk'",
+    "clean": "npx react-native-clean-project",
+    "clean-all": "npx react-native clean-project-auto",
+    "expo:start": "expo start",
+    "expo:android": "expo start --android",
+    "expo:ios": "expo start --ios",
+    "expo:web": "expo start --web",
+    "expo:build:detox": "detox build -c ios.sim.expo",
+    "expo:test:detox": "./bin/downloadExpoApp.sh && detox test --configuration ios.sim.expo"
+  },
+  "overrides": {
+    "react-error-overlay": "6.0.9"
   },
   "dependencies": {
-    "@react-native-community/async-storage": "^1.12.1",
-    "@react-native-community/masked-view": "0.1.10",
-    "@react-navigation/native": "^5.8.10",
-    "@react-navigation/stack": "^5.12.8",
-    "@unimodules/core": "^6.0.0",
-    "apisauce": "2.0.0",
-    "expo": "~39.0.2",
-    "expo-localization": "^9.1.0",
-    "expo-status-bar": "~1.0.2",
-    "i18n-js": "^3.8.0",
-    "mobx": "^6.0.4",
-    "mobx-react-lite": "^3.1.6",
-    "mobx-state-tree": "4.0.2",
-    "ramda": "0.27.1",
-    "react": "16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.2.tar.gz",
-    "react-native-gesture-handler": "1.8.0",
-    "react-native-keychain": "6.2.0",
-    "react-native-safe-area-context": "3.1.4",
-    "react-native-screens": "~2.13.0",
-    "react-native-splash-screen": "3.2.0",
-    "react-native-unimodules": "^0.11.0",
-    "reactotron-mst": "^3.1.3",
-    "reactotron-react-native": "^5.0.0",
-    "validate.js": "0.13.1"
+    "@expo-google-fonts/space-grotesk": "^0.2.2",
+    "@expo/webpack-config": "^0.17.0",
+    "@react-native-async-storage/async-storage": "^1.17.3",
+    "@react-navigation/bottom-tabs": "^6.3.2",
+    "@react-navigation/native": "~6.0.1",
+    "@react-navigation/native-stack": "^6.0.2",
+    "@react-navigation/stack": "~6.2.1",
+    "apisauce": "2.1.5",
+    "date-fns": "^2.29.2",
+    "expo": "~46.0.3",
+    "expo-application": "~4.2.2",
+    "expo-constants": "~13.2.3",
+    "expo-device": "~4.3.0",
+    "expo-file-system": "~14.1.0",
+    "expo-font": "~10.2.0",
+    "expo-linear-gradient": "~11.4.0",
+    "expo-localization": "13.1.0",
+    "expo-modules-core": "~0.11.3",
+    "expo-splash-screen": "~0.16.1",
+    "expo-status-bar": "~1.4.0",
+    "i18n-js": "3.9.2",
+    "mobx": "6.6.0",
+    "mobx-react-lite": "3.4.0",
+    "mobx-state-tree": "5.1.5",
+    "react": "18.0.0",
+    "react-native": "0.69.5",
+    "react-native-bootsplash": "4.3.2",
+    "react-native-gesture-handler": "~2.5.0",
+    "react-native-reanimated": "~2.9.1",
+    "react-native-safe-area-context": "4.3.4",
+    "react-native-screens": "3.15.0",
+    "reactotron-mst": "3.1.4",
+    "reactotron-react-js": "^3.3.7",
+    "reactotron-react-native": "5.0.3"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-decorators": "^7.12.1",
-    "@babel/plugin-proposal-optional-catch-binding": "^7.12.1",
-    "@storybook/addon-storyshots": "6.1.10",
-    "@storybook/react-native": "5.3.23",
-    "@storybook/react-native-server": "5.3.23",
-    "@types/i18n-js": "^3.0.3",
-    "@types/jest": "^26.0.18",
-    "@types/ramda": "0.27.32",
-    "@types/react": "~16.9.35",
-    "@types/react-dom": "~16.9.8",
-    "@types/react-native": "~0.63.2",
-    "@types/react-test-renderer": "^16.9.4",
-    "@typescript-eslint/eslint-plugin": "^4.10.0",
-    "@typescript-eslint/parser": "^4.10.0",
-    "babel-jest": "^26.6.3",
-    "babel-loader": "8.2.2",
-    "bufferutil": "4.0.2",
-    "canvas": "2.6.1",
-    "detox": "^17.6.1",
-    "detox-expo-helpers": "^0.6.0",
-    "eslint": "^7.15.0",
-    "eslint-config-prettier": "^7.0.0",
-    "eslint-config-standard": "^16.0.2",
-    "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "eslint-plugin-react": "^7.21.5",
-    "eslint-plugin-react-native": "^3.10.0",
-    "expo-detox-hook": "^1.0.10",
-    "jest": "25.5.4",
-    "jest-circus": "25.5.4",
-    "jest-expo": "39.0.0",
-    "jetifier": "^1.6.2",
-    "mocha": "8.2.1",
-    "npm-run-all": "4.1.5",
-    "patch-package": "6.2.2",
+    "@babel/core": "^7.18.0",
+    "@babel/plugin-proposal-decorators": "7.18.2",
+    "@babel/plugin-proposal-optional-catch-binding": "7.16.7",
+    "@babel/preset-env": "^7.16.11",
+    "@babel/runtime": "^7.18.3",
+    "@react-native-community/cli-platform-ios": "^8.0.2",
+    "@rnx-kit/metro-config": "^1.2.32",
+    "@rnx-kit/metro-resolver-symlinks": "0.1.21",
+    "@types/i18n-js": "3.8.2",
+    "@types/jest": "26.0.19",
+    "@types/react": "^17",
+    "@types/react-native": "0.67.8",
+    "@typescript-eslint/eslint-plugin": "5.28.0",
+    "@typescript-eslint/parser": "5.28.0",
+    "babel-jest": "26.6.3",
+    "babel-loader": "8.2.5",
+    "detox": "19.12.1",
+    "detox-expo-helpers": "0.6.0",
+    "eslint": "8.17.0",
+    "eslint-config-prettier": "8.5.0",
+    "eslint-config-standard": "17.0.0",
+    "eslint-plugin-import": "2.26.0",
+    "eslint-plugin-n": "^15.0.0",
+    "eslint-plugin-node": "11.1.0",
+    "eslint-plugin-promise": "6.0.0",
+    "eslint-plugin-react": "7.30.0",
+    "eslint-plugin-react-native": "4.0.0",
+    "expo-detox-hook": "1.0.10",
+    "expo-modules-autolinking": "0.10.1",
+    "fbjs-scripts": "3.0.1",
+    "jest": "26",
+    "jest-circus": "26",
+    "jest-environment-node": "26",
+    "jest-expo": "^46.0.1",
+    "metro-config": "0.71.1",
+    "metro-react-native-babel-preset": "0.71.1",
+    "metro-source-map": "0.71.1",
+    "mocha": "6",
+    "patch-package": "6.4.7",
     "postinstall-prepare": "1.0.1",
-    "prettier": "^2.2.1",
-    "react-devtools-core": "4.10.1",
-    "react-dom": "16.14.0",
-    "react-native-web": "~0.13.12",
-    "react-powerplug": "1.0.0",
-    "react-test-renderer": "^17.0.1",
-    "rimraf": "3.0.2",
-    "solidarity": "3.0.0",
-    "typescript": "~3.9.5",
-    "utf-8-validate": "5.0.3"
+    "prettier": "2.6.2",
+    "query-string": "^7.0.1",
+    "react-devtools-core": "4.24.7",
+    "react-dom": "18.0.0",
+    "react-native-web": "^0.18.7",
+    "reactotron-core-client": "^2.8.10",
+    "regenerator-runtime": "^0.13.4",
+    "ts-jest": "26",
+    "typescript": "4.7.3",
+    "webpack": "4"
   },
-  "jest": {
-    "preset": "jest-expo",
-    "setupFiles": [
-      "<rootDir>/node_modules/react-native/jest/setup.js",
-      "<rootDir>/test/setup.ts"
-    ],
-    "testPathIgnorePatterns": [
-      "/node_modules/",
-      "/e2e"
-    ],
-    "transformIgnorePatterns": [
-      "node_modules/(?!(jest-)?react-native|@react-navigation|@storybook|@react-native-community|expo-localization|@unimodules)"
-    ],
-    "projects": [
-      {
-        "preset": "jest-expo/ios",
-        "transformIgnorePatterns": [
-          "node_modules/(?!(jest-)?react-native|react-clone-referenced-element|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|native-base|@storybook)"
-        ],
-        "testPathIgnorePatterns": [
-          "/node_modules/",
-          "/e2e"
-        ]
-      },
-      {
-        "preset": "jest-expo/android",
-        "transformIgnorePatterns": [
-          "node_modules/(?!(jest-)?react-native|react-clone-referenced-element|@react-native-community|expo(nent)?|@expo(nent)?/.*|react-navigation|@react-navigation/.*|@unimodules/.*|native-base|@storybook)"
-        ],
-        "testPathIgnorePatterns": [
-          "/node_modules/",
-          "/e2e"
-        ]
-      }
-    ]
+  "resolutions": {
+    "@types/react": "^17",
+    "@types/react-dom": "^17"
   },
   "prettier": {
     "printWidth": 100,
