@@ -1,5 +1,5 @@
-const en = require("../app/i18n/en.json")
-const { exec } = require("child_process")
+import en from "../app/i18n/en"
+import { exec } from "child_process"
 
 // Use this array for keys that for whatever reason aren't greppable so they
 // don't hold your test suite hostage by always failing.
@@ -27,10 +27,12 @@ function iterate(obj, stack, array) {
  * It was taken from https://gist.github.com/Michaelvilleneuve/8808ba2775536665d95b7577c9d8d5a1
  * and modified slightly to account for our Ignite higher order components,
  * which take 'tx' and 'fooTx' props.
- * The grep command is nasty looking, but it's essentially searching the codebase for 3 things:
+ * The grep command is nasty looking, but it's essentially searching the codebase for a few different things:
  *
  * tx="*"
  * Tx=""
+ * tx={""}
+ * Tx={""}
  * translate(""
  *
  * and then grabs the i18n key between the double quotes
@@ -43,8 +45,8 @@ function iterate(obj, stack, array) {
 describe("i18n", () => {
   test("There are no missing keys", (done) => {
     // Actual command output:
-    // grep "Tx=\"\S*\"\|tx=\"\S*\"\|translate(\"\S*\"" -ohr './app' | grep -o "\".*\""
-    const command = `grep "Tx=\\"\\S*\\"\\|tx=\\"\\S*\\"\\|translate(\\"\\S*\\"" -ohr './app' | grep -o "\\".*\\""`
+    // grep "[T\|t]x=[{]\?\"\S*\"[}]\?\|translate(\"\S*\"" -ohr './app' | grep -o "\".*\""
+    const command = `grep "[T\\|t]x=[{]\\?\\"\\S*\\"[}]\\?\\|translate(\\"\\S*\\"" -ohr './app' | grep -o "\\".*\\""`
     exec(command, (_, stdout) => {
       const allTranslationsDefined = iterate(en, "", [])
       const allTranslationsUsed = stdout.replace(/"/g, "").split("\n")
