@@ -1,14 +1,15 @@
-// This is the first file that ReactNative will run when it starts up.
-//
-// We jump out of here immediately and into our main entry point instead.
-//
-// It is possible to have React Native load our main module first, but we'd have to
-// change that in both AppDelegate.m and MainApplication.java.  This would have the
-// side effect of breaking other tooling like mobile-center and react-native-rename.
-//
-// It's easier just to leave it here.
+// This is the entry point if you run `yarn expo:start`
+// If you run `yarn ios` or `yarn android`, it'll use ./index.js instead.
 import App from "./app/app.tsx"
-import { AppRegistry } from "react-native"
+import React from "react"
+import { registerRootComponent } from "expo"
+import * as SplashScreen from "expo-splash-screen"
 
-AppRegistry.registerComponent("PizzaApp", () => App)
-export default App
+SplashScreen.preventAutoHideAsync()
+
+function IgniteApp() {
+  return <App hideSplashScreen={SplashScreen.hideAsync} />
+}
+
+registerRootComponent(IgniteApp)
+export default IgniteApp
