@@ -1,10 +1,30 @@
 import i18n from "i18n-js"
+import { TxKeyPath } from "./i18n"
 
 /**
  * Translates text.
  *
  * @param key The i18n key.
+ * @param options The i18n options.
+ * @returns The translated text.
+ *
+ * @example
+ * Translations:
+ *
+ * ```en.ts
+ * {
+ *  "hello": "Hello, {{name}}!"
+ * }
+ * ```
+ *
+ * Usage:
+ * ```ts
+ * import { translate } from "i18n-js"
+ *
+ * translate("common.ok", { name: "world" })
+ * // => "Hello world!"
+ * ```
  */
-export function translate(key: string, options?: I18n.TranslateOptions) {
-  return key ? i18n.t(key, options) : null
+export function translate(key: TxKeyPath, options?: i18n.TranslateOptions) {
+  return i18n.t(key, options)
 }
