@@ -1,6 +1,4 @@
-export * from "./extensions/with-environment"
-export * from "./extensions/with-root-store"
-export * from "./extensions/with-status"
-export * from "./root-store/root-store"
-export * from "./root-store/root-store-context"
-export * from "./root-store/setup-root-store"
+export * from "./RootStore"
+export * from "./helpers/getRootStore"
+export * from "./helpers/useStores"
+export * from "./helpers/setupRootStore"
