@@ -1,4 +1,4 @@
-export * from "./color"
+export * from "./colors"
 export * from "./spacing"
 export * from "./typography"
 export * from "./timing"
